+++ test/config_test.rb
@@ -10,8 +10,7 @@ module Judoscale
         config = Config.instance
         _(config.api_base_url).must_equal "https://example.com"
         _(config.dyno).must_equal "web.1"
-        _(config.debug).must_equal false
-        _(config.quiet).must_equal false
+        _(config.log_level).must_be_nil
         _(config.logger).must_equal Rails.logger
         _(config.max_request_size_bytes).must_equal 100_000
         _(config.report_interval_seconds).must_equal 10
@@ -29,14 +28,14 @@ module Judoscale
       env = {
         "DYNO" => "web.2",
         "JUDOSCALE_URL" => "https://custom.example.com",
-        "JUDOSCALE_DEBUG" => "true"
+        "JUDOSCALE_LOG_LEVEL" => "debug"
       }
 
       use_env env do
         config = Config.instance
         _(config.api_base_url).must_equal "https://custom.example.com"
         _(config.dyno).must_equal "web.2"
-        _(config.debug).must_equal true
+        _(config.log_level).must_equal ::Logger::Severity::DEBUG
       end
     end
 
@@ -46,8 +45,7 @@ module Judoscale
       Judoscale.configure do |config|
         config.dyno = "web.3"
         config.api_base_url = "https://block.example.com"
-        config.debug = true
-        config.quiet = true
+        config.log_level = :info
         config.logger = test_logger
         config.max_request_size_bytes = 50_000
         config.report_interval_seconds = 20
@@ -59,8 +57,7 @@ module Judoscale
       config = Config.instance
       _(config.api_base_url).must_equal "https://block.example.com"
       _(config.dyno).must_equal "web.3"
-      _(config.debug).must_equal true
-      _(config.quiet).must_equal true
+      _(config.log_level).must_equal ::Logger::Severity::INFO
       _(config.logger).must_equal test_logger
       _(config.max_request_size_bytes).must_equal 50_000
       _(config.report_interval_seconds).must_equal 20