+++ test/worker_adapters/que_test.rb
@@ -45,7 +45,7 @@ module Judoscale
       end
 
       it "logs debug information for each queue being collected" do
-        use_config debug: true do
+        use_config log_level: :debug do
           enqueue("default", Time.now)
 
           subject.collect! store